+++ objectified-rest/tests/test_toolchain_packaging.py
@@ -42,6 +42,7 @@ _EXPECTED_KEYS = {
     "asyncapi-parser",
     "asyncapi-diff",
     "rover",
+    "graphql-inspector-diff",
 }
 
 _AUTH = {